make withdrawals page top level
